+++ messages/fr/base.php
@@ -1,8 +1,14 @@
 <?php
-return array (
-  'Content not available' => 'Le contenu n\'est pas disponible',
-  'Share' => 'Partager',
-  'Shared content' => 'Partager ce contenu',
-  'This content has either been deleted or you no longer have permission to access it.' => 'Ce contenu a été supprimé ou vous n\'êtes plus autorisé à y accéder.',
-  '{spaceName} by {userName}' => '{spaceName} par {userName}',
-);
+
+return [
+    '<strong>Share</strong> content' => '',
+    'Content not available' => '',
+    'Content you create is automatically displayed on your profile.' => '',
+    'Select Spaces here on which the content is to be additionally displayed.' => '',
+    'Share' => '',
+    'Share this content on your profile stream' => '',
+    'Shared content' => '',
+    'Spaces' => '',
+    'This content has either been deleted or you no longer have permission to access it.' => '',
+    '{spaceName} by {userName}' => '',
+];