+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 1.0.3 (Unreleased)
 -------------------
 - Fix #39: Deny to share a not published content
+- Fix #45: Disable possibility to share on the profile stream when "profileDisableStream" is true
 
 1.0.2 (May 1, 2023)
 -------------------